+++ lib/vulkan/renpass/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectory(atmos)
 add_subdirectory(bloom)
 add_subdirectory(compo)
+add_subdirectory(envmap)
