+++ lib/vulkan/renpass/envmap/src/mip_chain.cpp
@@ -1,10 +1,10 @@
-#include "renderpass/envmap/envmap.hpp"
+#include "mirinae/vulkan/renpass/envmap/envmap.hpp"
 
 #include "mirinae/lightweight/task.hpp"
+#include "mirinae/vulkan/base/render/cmdbuf.hpp"
+#include "mirinae/vulkan/base/renderpass/builder.hpp"
 
-#include "render/cmdbuf.hpp"
-#include "renderpass/builder.hpp"
-#include "renderpass/envmap/cubemap.hpp"
+#include "cubemap.hpp"
 
 
 // Tasks
