+++ lib/vulkan/renpass/envmap/src/rp_base.cpp
@@ -1,16 +1,21 @@
-#include "renderpass/envmap/rp.hpp"
+#include "rp.hpp"
 
-#include "renderpass/builder.hpp"
+#include "mirinae/vulkan/base/renderpass/builder.hpp"
 
 
-namespace {
+// env_base
+namespace { namespace env_base {
 
-    VkDescriptorSetLayout create_desclayout_main(
+    VkDescriptorSetLayout create_desclayout_model(
         mirinae::DesclayoutManager& desclayouts, mirinae::VulkanDevice& device
     ) {
-        mirinae::DescLayoutBuilder builder{ "env_sky:main" };
-        builder.add_img(VK_SHADER_STAGE_FRAGMENT_BIT, 1);  // Equirectangular
-        return desclayouts.add(builder, device.logi_device());
+        return desclayouts.get("gbuf:model").layout();
+    }
+
+    VkDescriptorSetLayout create_desclayout_actor(
+        mirinae::DesclayoutManager& desclayouts, mirinae::VulkanDevice& device
+    ) {
+        return desclayouts.get("gbuf:actor").layout();
     }
 
     VkRenderPass create_renderpass(
@@ -21,12 +26,13 @@ namespace {
         builder.attach_desc()
             .add(depth)
             .ini_layout(VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL)
-            .fin_layout(VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL);
+            .fin_layout(VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL)
+            .op_pair_clear_store();
         builder.attach_desc()
             .add(color)
-            .ini_layout(VK_IMAGE_LAYOUT_UNDEFINED)
+            .ini_layout(VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL)
             .fin_layout(VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL)
-            .op_pair_clear_store();
+            .op_pair_load_store();
 
         builder.color_attach_ref().add_color_attach(1);
 
@@ -45,8 +51,16 @@ namespace {
         mirinae::PipelineBuilder builder{ device };
 
         builder.shader_stages()
-            .add_vert(":asset/spv/env_sky_vert.spv")
-            .add_frag(":asset/spv/env_sky_frag.spv");
+            .add_vert(":asset/spv/env_base_vert.spv")
+            .add_frag(":asset/spv/env_base_frag.spv");
+
+        builder.vertex_input_state().set_static();
+
+        builder.rasterization_state().cull_mode(VK_CULL_MODE_FRONT_BIT);
+
+        builder.depth_stencil_state()
+            .depth_test_enable(true)
+            .depth_write_enable(true);
 
         builder.color_blend_state().add(false, 1);
 
@@ -76,9 +90,10 @@ namespace {
                 formats_.at(0), formats_.at(1), device.logi_device()
             );
             layout_ = mirinae::PipelineLayoutBuilder{}
-                          .desc(create_desclayout_main(desclayouts, device))
+                          .desc(create_desclayout_model(desclayouts, device))
+                          .desc(create_desclayout_actor(desclayouts, device))
                           .add_vertex_flag()
-                          .pc<mirinae::U_EnvSkyPushConst>()
+                          .pc<mirinae::U_EnvmapPushConst>()
                           .build(device);
             pipeline_ = create_pipeline(renderpass_, layout_, device);
         }
@@ -108,15 +123,15 @@ namespace {
         std::array<VkClearValue, 2> clear_values_;
     };
 
-}  // namespace
+}}  // namespace ::env_base
 
 
 namespace mirinae {
 
-    std::unique_ptr<IRenPass> create_rp_sky(
+    std::unique_ptr<IRenPass> create_rp_base(
         DesclayoutManager& desclayouts, VulkanDevice& device
     ) {
-        return std::make_unique<::RPBundle>(desclayouts, device);
+        return std::make_unique<env_base::RPBundle>(desclayouts, device);
     }
 
 }  // namespace mirinae
