+++ lib/vulkan/renpass/envmap/src/rp_brdf_lut.cpp
@@ -1,6 +1,6 @@
-#include "renderpass/envmap/rp.hpp"
+#include "rp.hpp"
 
-#include "renderpass/builder.hpp"
+#include "mirinae/vulkan/base/renderpass/builder.hpp"
 
 
 namespace {
