+++ lib/vulkan/renpass/envmap/src/rp_diffuse.cpp
@@ -1,6 +1,6 @@
-#include "renderpass/envmap/rp.hpp"
+#include "rp.hpp"
 
-#include "renderpass/builder.hpp"
+#include "mirinae/vulkan/base/renderpass/builder.hpp"
 
 
 namespace {
@@ -8,7 +8,9 @@ namespace {
     VkDescriptorSetLayout create_desclayout_main(
         mirinae::DesclayoutManager& desclayouts, mirinae::VulkanDevice& device
     ) {
-        return desclayouts.get("envdiffuse:main").layout();
+        mirinae::DescLayoutBuilder builder{ "envdiffuse:main" };
+        builder.add_img(VK_SHADER_STAGE_FRAGMENT_BIT, 1);  // envmap
+        return desclayouts.add(builder, device.logi_device());
     }
 
     VkRenderPass create_renderpass(VkFormat color, VkDevice logi_device) {
@@ -35,8 +37,8 @@ namespace {
         mirinae::PipelineBuilder builder{ device };
 
         builder.shader_stages()
-            .add_vert(":asset/spv/env_specular_vert.spv")
-            .add_frag(":asset/spv/env_specular_frag.spv");
+            .add_vert(":asset/spv/env_diffuse_vert.spv")
+            .add_frag(":asset/spv/env_diffuse_frag.spv");
 
         builder.color_blend_state().add(false, 1);
 
@@ -64,8 +66,7 @@ namespace {
             layout_ = mirinae::PipelineLayoutBuilder{}
                           .desc(create_desclayout_main(desclayouts, device))
                           .add_vertex_flag()
-                          .add_frag_flag()
-                          .pc<mirinae::U_EnvSpecularPushConst>()
+                          .pc<mirinae::U_EnvdiffusePushConst>()
                           .build(device);
             pipeline_ = create_pipeline(renderpass_, layout_, device);
         }
@@ -100,7 +101,7 @@ namespace {
 
 namespace mirinae {
 
-    std::unique_ptr<IRenPass> create_rp_specular(
+    std::unique_ptr<IRenPass> create_rp_diffuse(
         DesclayoutManager& desclayouts, VulkanDevice& device
     ) {
         return std::make_unique<::RPBundle>(desclayouts, device);
